% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_Toeplitz.R
\name{AR1_mat}
\alias{AR1_mat}
\title{Generate AR1 structure matrix}
\usage{
AR1_mat(rho, p, band = NULL)
}
\arguments{
\item{rho}{a numeric value, \eqn{\rho} in AR structure}

\item{p}{a integer, number of column of matrix}

\item{band=NULL}{a integer, banded AR(1) covariance matrix, default is not banded.}
}
\description{
Generate AR1 structure matrix
}
