% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoRegCoef_design_Toeplitz.R
\name{autoRegCoef_design_Toeplitz}
\alias{autoRegCoef_design_Toeplitz}
\title{Generate a toeplitz cholesky factor design matrix}
\usage{
autoRegCoef_design_Toeplitz(d)
}
\arguments{
\item{d}{The dimension of the matrix}
}
\value{
The design matrix
}
\description{
Generate design matrix for W, the autoregression coefficient model design matrix in Joint Mean Covariance Model
The result is a d*(d-1)/2*d design matrix, multiply the parameter coefficient lambda, then can have a wide expand vector,
use BayesJMCM2::vec2mat can transfer back to a lower-triangular matrix,
}
\examples{
BayesJMCM2::vector2ltril(autoRegCoef_design_Toeplitz(5)\%*\% c(4,3,2,1))
\dontrun{vector2ltril(autoRegCoef_design_Toeplitz(5)\%*\% c(4,3,2,1))}
}
