% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_Toeplitz.R
\name{ToeplitzMat}
\alias{ToeplitzMat}
\title{Generate Toeplitz structure matrix given first row of matrix}
\usage{
ToeplitzMat(toeplitz)
}
\arguments{
\item{toepliz}{A vector correspond the first row of Toeplitz matrix}
}
\value{
The corresponding toeplitz matrix
}
\description{
Generate Toeplitz structure matrix given first row of matrix
}
\examples{
ToeplitzMat(c(1,2,3))
ToeplitzMat(c(1,0.5,0.5))
ToeplitzMat(cumprod(c(1,rep(0.7,5))))
ToeplitzMat(c(1,1,0,0))
}
